feat(browser-ext): extract wait/pagination helpers to extraction-pipeline.ts (Phase 3e batch 1)
Create createExtractionPipeline(deps) DI factory in lib/extraction-pipeline.ts.
Extract 7 inline functions from content.ts:
- isDisabledPaginationControl
- waitForResumeCards, waitForApiRows, waitForExtractionData
- clearCapturedResultsForNextPage, waitForSeekProfileSnapshot
Guard functions (loadCollectionGuards, parseGuardFieldNames, applyCollectionGuards)
and isElementVisible remain inline — referenced by extractor factory calls at
module init before pipeline factory call executes.

Content.ts: 5,316L → 5,176L (-140L). Pipeline factory: +266L.
make check: 0 errors.
